requirements.txt
python: 3.10
pytorch: 2.5.0
torchvision: 0.20.0

timm
absl-py==2.1.0
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==24.3.0
Brotli==1.0.9
certifi==2024.12.14
charset-normalizer==2.0.4
clean-fid==0.1.35
click==8.1.8
clip-anytorch==2.6.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
DateTime==5.5
dctorch==0.1.2
docker-pycreds==0.4.0
einops==0.8.0
fairscale==0.4.13
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.3.1
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.1.2
grpcio==1.68.1
huggingface-hub==0.26.5
idna==3.4
imageio==2.36.1
Jinja2==3.1.4
jsonmerge==1.9.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
k-diffusion==0.1.1.post1
kiwisolver==1.4.5
kornia==0.7.4
kornia_rs==0.1.8
lazy_loader==0.4
lightning-utilities==0.11.9
Markdown==3.7
MarkupSafe==2.1.3
matplotlib==3.9.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.0
pillow==10.2.0
pip==23.3.1
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.1
pydantic==2.10.4
pydantic_core==2.27.2
pyDeprecate==0.3.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-lightning==1.4.2
pytz==2024.2
PyYAML==6.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
rpds-py==0.22.3
safetensors==0.4.3
scikit-image==0.25.0
scipy==1.14.1
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==68.2.2
six==1.16.0
smmap==5.0.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tifffile==2024.12.12
timm==0.9.16
tokenizers==0.21.0
torch==2.5.0
torchaudio==2.5.0
torchdiffeq==0.2.5
torchmetrics==0.6.0
torchsde==0.2.6
torchvision==0.20.0
tqdm==4.66.2
trampoline==0.1.2
transformers==4.47.0
typing_extensions==4.12.2
urllib3==2.1.0
wandb==0.19.1
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.41.2
win-inet-pton==1.1.0
yarl==1.18.3
zope.interface==7.2
